pyproject.toml
[tool.poetry]
name = "project-0"
version = "0.1.0"
description = ""
authors = ["Kostyantin Zahorodnyi <constantine2903@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
marshmallow = "^3.23.2"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
colorlog = "^6.9.0"
celery = {extras = ["redis"], version = "^5.4.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
